requirements.txt
PyJWT~=2.3.0
Flask~=2.0.2
SQLAlchemy~=1.4.35
Werkzeug~=2.0.2
Flask-RESTful~=0.3.9
waitress~=2.1.1